perf(unified): intern column keys as Arc<str> in UnifiedRecord
Replace HashMap<String, Value> with HashMap<Arc<str>, Value> across
the scan and join hot paths. Per-record key allocations drop to
Arc::clone; concurrent scenario closes from ~10x gap to ~2x on
mini-duel (473 -> 1157 ops/s wire).

Step 1 of roadmap #1 (UnifiedRecord layout). Step 2 — schema-shared
Vec<Value> — still pending and compounds for scan paths (select_range,
select_filtered) where Arc<str> alone is not enough.

Also fixes two unrelated test-only breaks exposed by -all-targets:
filter_compiled row_id u32->u64 and auth::AuthConfig missing cert/oauth.

+++ src/runtime/query_exec/filter_compiled.rs
@@ -1103,7 +1103,7 @@ mod tests {
     fn make_row_entity(id: u64, columns: Vec<(String, Value)>) -> UnifiedEntity {
         let kind = crate::storage::EntityKind::TableRow {
             table: std::sync::Arc::from("users"),
-            row_id: id as u32,
+            row_id: id,
         };
         let cols: Vec<Value> = columns.iter().map(|(_, v)| v.clone()).collect();
         let names: Vec<String> = columns.iter().map(|(k, _)| k.clone()).collect();
